requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.1.0
fastapi==0.70.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
motor==2.5.1
Pillow==9.0.0
pydantic==1.9.0
pymongo==3.12.3
requests==2.27.1
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
